+++ R/time_sample.R
@@ -14,7 +14,7 @@ as_time <- function(x, .unit = "second") {
 
 #' @export
 as_time.sample_int <- function(x, .unit = "second") {
-  time <- (x - 1) / scaling(sampling_rate = attributes(x)$sampling_rate, .unit)
+  time <- (x - 1) / scaling(attributes(x)$sampling_rate, .unit)
   attributes(time) <- NULL
   time
 }
@@ -36,16 +36,16 @@ as_sample_int <- function(x, ...) {
 #' @rdname as_sample_int
 #' @param x A vector of numeric values.
 #' @param .unit "seconds" (or "s"), "milliseconds" (or "ms"), or "samples"
-#' @param sampling_rate Sampling rate in Hz
+#' @param .sampling_rate Sampling rate in Hz
 #' @export
-as_sample_int.numeric <- function(x, sampling_rate = NULL, .unit = "s", ...) {
-  if (is.null(sampling_rate)) stop("'sampling_rate' needs to be specified", call. = FALSE)
-  scale <- scaling(sampling_rate, unit = .unit)
+as_sample_int.numeric <- function(x, .sampling_rate = NULL, .unit = "s", ...) {
+  if (is.null(.sampling_rate)) stop("'.sampling_rate' needs to be specified", call. = FALSE)
+  scale <- scaling(.sampling_rate, unit = .unit)
   # shift if it's in time scale so that sample 1 corresponds to time 0,
   # but not shift if it's converting samples into samples
   shift <- if (scale == 1) 0 else 1
   samples <- round(x * scale + shift)
-  sample_int(samples, sampling_rate)
+  sample_int(samples, .sampling_rate)
 }
 #' @export
 as_sample_int.sample_int <- function(x, ...) {