+++ tests/testthat/test_05-tv_05-group_by.R
@@ -9,7 +9,7 @@ data_1 <- eeguana:::data_sincos3id
 #     dplyr::tibble(
 #       X = sin(1:30), Y = cos(1:30),
 #       .id = rep(c(1L, 2L, 3L), each = 10),
-#       .sample = sample_int(rep(seq(-4L, 5L), times = 3), sampling_rate = 500)
+#       .sample = sample_int(rep(seq(-4L, 5L), times = 3), .sampling_rate = 500)
 #     ),
 #   channels_tbl = dplyr::tibble(
 #     .channel = c("X", "Y"), .reference = NA, theta = NA, phi = NA,