+++ tests/testthat/test_12-constr.R
@@ -32,23 +32,23 @@ test_that("[[ works", {
 test_that("eeg_lst", {
   
 N<- 1000
-df0 <- eeg_lst(signal_tbl = data.frame(.id=1L, .sample = sample_int(1:N, sampling_rate = 500),
+df0 <- eeg_lst(signal_tbl = data.frame(.id=1L, .sample = sample_int(1:N, .sampling_rate = 500),
                                       X = channel_dbl(sin(seq(0, 8*pi, length.out = N)))))
-df1 <- eeg_lst(signal_tbl = data.frame(.sample = sample_int(1:N, sampling_rate = 500),
+df1 <- eeg_lst(signal_tbl = data.frame(.sample = sample_int(1:N, .sampling_rate = 500),
                                        X = channel_dbl(sin(seq(0, 8*pi, length.out = N)))))
 df2 <- eeg_lst(signal_tbl = data.frame(.id=1L, X = channel_dbl(sin(seq(0, 8*pi, length.out = N)))),
-              sampling_rate = 500)
+              .sampling_rate = 500)
 
 df3 <- eeg_lst(signal_tbl = data.frame(X = channel_dbl(sin(seq(0, 8*pi, length.out = N)))),
-      sampling_rate = 500)
+      .sampling_rate = 500)
 df4 <- eeg_lst(signal_tbl = data.table::data.table(X = channel_dbl(sin(seq(0, 8*pi, length.out = N)))),
-               sampling_rate = 500)
-df5 <- eeg_lst(signal_tbl = data.table::data.table(.id=1L, .sample = sample_int(1:N, sampling_rate = 500),
+               .sampling_rate = 500)
+df5 <- eeg_lst(signal_tbl = data.table::data.table(.id=1L, .sample = sample_int(1:N, .sampling_rate = 500),
                                        X = channel_dbl(sin(seq(0, 8*pi, length.out = N)))))
 
 df6 <- eeg_lst(signal_tbl = dplyr::tibble(X = channel_dbl(sin(seq(0, 8*pi, length.out = N)))),
-               sampling_rate = 500)
-df7 <- eeg_lst(signal_tbl =dplyr::tibble(.id=1L, .sample = sample_int(1:N, sampling_rate = 500),
+               .sampling_rate = 500)
+df7 <- eeg_lst(signal_tbl =dplyr::tibble(.id=1L, .sample = sample_int(1:N, .sampling_rate = 500),
                                                    X = channel_dbl(sin(seq(0, 8*pi, length.out = N)))))
 expect_equal(df0,df1)
 expect_equal(df0,df2)
@@ -59,3 +59,4 @@ expect_equal(df0,df6)
 expect_equal(df0,df7)
 
 })
+